app: Make format function for time strings more generic
We can use the same format function for different types of "Last Seen",
"Last accessed" and "Last online".
